+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Types of changes are:
 * Fixed bug causing some properties to be parsed twice.
 * Properties which conflict with Python keywords are now re-mapped
   on generated models.
+* Fixed bug where default is overriden on Object classes in parsers.
 
 ## [0.4.0] - 2020-03-29
 ### Added